Generate function name and C arg types from subroutine/entry definition

Parse a fortran subroutine/entry definition and return subroutine/entry name and C argument types as a list

make_fun_list(fun_strs)

Arguments

fun_str
a vector of subroutine definition strings with args

Value

a list of function names along with C argument types

Examples

make_fun_list(c("subroutine foo(no,ni,x,y,w,theta)", "entry foobar(ix, y)", "subroutine bar(x,n,p,m)"))

#> [[1]]
#> [[1]]$name
#> [1] "foo"
#> 
#> [[1]]$args
#> [1] "int *no"       "int *ni"       "double *x"     "double *y"    
#> [5] "double *w"     "double *theta"
#> 
#> 
#> [[2]]
#> [[2]]$name
#> [1] "foobar"
#> 
#> [[2]]$args
#> [1] "int *ix"   "double *y"
#> 
#> 
#> [[3]]
#> [[3]]$name
#> [1] "bar"
#> 
#> [[3]]$args
#> [1] "double *x" "int *n"    "double *p" "int *m"   
#> 
#>